added two new core motion examples.
