+++ examples/ios/iosCoreMotionLegacyExample/src/main.mm
